GBZ80: Force the RHS of 8-bit arith ops to not be A.
We can get into a situation with a GPR on the left
and actual A on the right, like this:

  %rc = ADD8r %rc, %ra

We can't expand this unless we commute, but we
can't commute SUB, SBC and CP. Just prevent the
RHS from being A for now.

XXX: This could be a problem for patterns that
are legal, like 'ADD A, A'.