GBZ80: Improve the call sequence a bit.
Now we insert pushes for each value in the
call sequence. Hopefully there is room for
optimizing this later.

This optimization will likely have to be part
of the calling convention itself somehow, since
both caller and callee need to know how the
stack values are 'packed'.

We also copy register parameters in reverse
order to get HL/A last. Hopefully this works
the way we want.